requirements.txt
pycoingecko==1.4.1
django==2.1.7  
django-cors-headers==2.5.2
djangorestframework==3.9.2 
requests==2.21.0
gunicorn==19.9.0
